+++ lib/modules/state/state_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: state
-// Updated at: 2026-05-16 22:44:54
-// Commit: perf(state): apply RepaintBoundary on animated state widgets
+// Updated at: 2026-05-17 22:46:31
+// Commit: feat(state): integrate custom telemetry logger in state
 
 class stateComponentState {
-  final String commitRef = '1217';
-  final String lastMessage = 'perf(state): apply RepaintBoundary on animated state widgets';
-  final DateTime timestamp = DateTime.parse('2026-05-16T22:44:54+05:00');
+  final String commitRef = '1229';
+  final String lastMessage = 'feat(state): integrate custom telemetry logger in state';
+  final DateTime timestamp = DateTime.parse('2026-05-17T22:46:31+05:00');
 }